New task

+++ 0x05-pointers_arrays_strings/2-strlen.c
@@ -7,7 +7,7 @@
  *Return: void thatmeas answer is correct
  */
 
-int_ _strlen(char *s)
+int _strlen(char *s)
 {
 	int u;
